chore(resource): remove get/set prefered encoding
This feature was removed for now as preferred encoding was not
part of the _shared_ resource but the resource handle so it
could lead to undetermenisem if there where multiple handles which
preferred different encoding (this was also why it was called
preferred encoding). It also is quite irrelevant for emails today,
e.g. it could be used to encode a attached json file with quoted
printable instead of base64, which is nice for a feature less
tui mail client, but doe they even still exist tody. Even current
tui mail clients should handle multipart mails.